periph: Change event handling API
Event handler has been changed to a base class with a specific function.
This is done so that drivers and threads can be made event handlers.

Peripherals now store pointers to event handlers in a vector. This
causes it to be dynamically allocated, something that needs to be
addressed.

The argument to event handling function now contains information about
the IRQ, peripheral where event occured (source) and optional data.